Synchronize cmt state with kernel evdev driver
A SYN_DROPPED event indicates the the kernel's internal event buffer
overflowed, and some events will not be delivered to userspace.
This can directly lead to 'ghost finger' (userspace thinking a finger
exists after it is no longer touching a pad), for example, if a
TRACKING_ID=-1 event is dropped for a given mt slot.

BUG=chromium-os:28997
TEST=run the following test on device and see if it is completed successfully.
  autotest platform_TrackpadSynDrop/control

+++ src/gesture.c
@@ -14,7 +14,7 @@
 #include "cmt.h"
 #include "properties.h"
 
-static unsigned MT_XButtons_To_Gestures_Buttons(unsigned);
+static unsigned MT_XButtons_To_Gestures_Buttons(unsigned long);
 
 /*
  * Gestures timer functions
@@ -132,7 +132,7 @@ Gesture_Device_Close(GesturePtr rec)
 }
 
 static unsigned
-MT_XButtons_To_Gestures_Buttons(unsigned xbuttons)
+MT_XButtons_To_Gestures_Buttons(unsigned long xbuttons)
 {
     unsigned ret = 0;
     if (xbuttons & BUTTON_LEFT)
